Remove deprecated classes and methods
